Refactor billing filter classes to use inheritance hierarchy
Introduce base filter classes `DateRangeFilter`, `PagedFilter`, and `BaseReport<T>` to eliminate code duplication. Move common date range properties to `DateRangeFilter`, pagination and ordering properties to `PagedFilter`. Extract currency-related properties into `CurrencyCode`, `CurrencyAmount`, and `CurrencyTotalAmount` base classes. Update `OperationFilter`, `MonthlyUsageFilter`, `UsageFilter`, `Report`, `UsageReport`, `CustomerMonthlyUsage`, `CustomerServiceUsage`, `TransactionInfo`, and `SubAccount` to inherit from appropriate base classes. Change `TotalQuantity` type from `int` to `long` in `BaseReport` and `PaymentsDto`.

+++ web/ASC.Web.Api/ApiModels/ResponseDto/PaymentsDto.cs
@@ -69,7 +69,7 @@ public class ReportDto
     /// The total quantity of operations in the report.
     /// </summary>
     /// <example>1</example>
-    public int TotalQuantity { get; set; }
+    public long TotalQuantity { get; set; }
     /// <summary>
     /// The total number of pages in the report.
     /// </summary>
